DELETE: micro_nutrient enum
We will not be using this , It is way too complicated to give
suggestions based on every micro there is.
The best way is probably to just eat a variety of vitamin rich fruit and
veggies and drink water with a lot of good minerals.
Again this could be changed in the future.